pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-envelope"
version = "0.1.0"
description = "Runtime behavioral envelope enforcement for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Kaustubh Phatak" }]
keywords = ["ai", "agents", "security", "mcp", "envelope", "behavioral-monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
agent-envelope = "agent_envelope.cli:main"

[project.urls]
Repository = "https://github.com/kphatak001/agent-envelope"
